bigquery/requirements.txt
mysql-replication
PyMySQL
google-cloud-bigquery